Pyno Interface 2.0
Added plenty of features including:
* Sending motor position commands to the robot
* Sending emergency stop command to the robot
* Receiving text messages from the robot (so far only "numerical"
messages were considered).

Improved the "auto-zoom" feature for the y_axis. Now it leaves a gap
above and below the data, which also prevents errors when all data has
the same y.

Added an emergency_stop function to stop the motor.

Added a check_limits function to verify that the robot is behaving
properly (so far just checks that m1_angle is below 90 degrees but a
cut-off current could be added here).